feat: add dual-backend model fitting (python/r) with backend parameter
Add a `backend` parameter to `ms_trawl()` that lets users choose between:

- backend="python" (default): pure Python using statsmodels WLS + Wald F-tests.
  No R dependency. Best for production/deployment and large experiments.

- backend="r": uses R's lme4 + pbkrtest via rpy2 for mixed-effects models
  and Kenward-Roger tests. Requires R + rpy2 + R packages. Best for
  small-sample experiments and publication-quality statistics.

Architecture:
- model_fitting.py: dispatcher that routes to python or r backend
- model_fitting_python.py: statsmodels WLS implementation (renamed from original)
- model_fitting_r.py: rpy2 bridge to R's adaptModel() from msTrawler package

The preprocessing layer (LOD, normalization, outliers, data reshaping) and
output layer (FDR correction, CSV writing) always run in pure Python regardless
of backend choice. Only the per-protein model fitting is swappable.

Also:
- pyproject.toml: optional [r] dependency group for rpy2
- check_backend() utility to verify availability before running
- README updated with backend comparison, architecture diagram, and guidance

13/13 E2E tests passing (python backend).
