+++ python/src/mstrawler.egg-info/SOURCES.txt
@@ -1,8 +1,11 @@
+README.md
 pyproject.toml
 src/mstrawler/__init__.py
 src/mstrawler/covariate_setup.py
 src/mstrawler/file_converter.py
 src/mstrawler/model_fitting.py
+src/mstrawler/model_fitting_python.py
+src/mstrawler/model_fitting_r.py
 src/mstrawler/pipeline.py
 src/mstrawler/preprocessing.py
 src/mstrawler/restructure.py
@@ -13,14 +16,21 @@ src/mstrawler.egg-info/dependency_links.txt
 src/mstrawler.egg-info/requires.txt
 src/mstrawler.egg-info/top_level.txt
 src/mstrawler/__pycache__/__init__.cpython-312.pyc
+src/mstrawler/__pycache__/covariate_setup.cpython-312.pyc
 src/mstrawler/__pycache__/file_converter.cpython-312.pyc
+src/mstrawler/__pycache__/model_fitting.cpython-312.pyc
+src/mstrawler/__pycache__/pipeline.cpython-312.pyc
 src/mstrawler/__pycache__/preprocessing.cpython-312.pyc
+src/mstrawler/__pycache__/restructure.cpython-312.pyc
+src/mstrawler/__pycache__/results.cpython-312.pyc
 tests/__init__.py
 tests/__pycache__/__init__.cpython-312.pyc
 tests/e2e/__init__.py
 tests/e2e/test_file_converter.py
+tests/e2e/test_pipeline.py
 tests/e2e/test_preprocessing.py
 tests/e2e/__pycache__/__init__.cpython-312.pyc
 tests/e2e/__pycache__/test_file_converter.cpython-312-pytest-9.0.2.pyc
+tests/e2e/__pycache__/test_pipeline.cpython-312-pytest-9.0.2.pyc
 tests/e2e/__pycache__/test_preprocessing.cpython-312-pytest-9.0.2.pyc
 tests/unit/__init__.py
